requirements.txt
imagehash
django
djangorestframework
pillow
yadisk
django-rest-knox
numpy
opencv-python
scikit-image
pytesseract
torch
torchvision
Google-Images-Search
pymorphy2
